requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pony==0.7.14
pycparser==2.20
PyPDF2==1.26.0
python-slugify==4.0.1
pythonnet==2.5.1
pywebview==3.4
text-unidecode==1.3
Werkzeug==1.0.1
xmltodict==0.12.0
